feat(components): docs-to-verify-list
