+++ src/views/admin/AdminDocuments.vue
@@ -23,6 +23,12 @@
           :label="ageGroups.find(group => group.value === age)?.label"
           @click="selectedAge = age"
         />
+        <Button
+          label="Сбросить"
+          icon="pi pi-refresh"
+          class="p-button-text p-button-sm"
+          @click="resetAge"
+        />
       </div>
 
       <!-- Фильтры -->
@@ -68,258 +74,23 @@
       </div>
     </div>
 
-    <div
-      v-if="filteredDocuments.filter((doc) => doc.verified !== false).length > 0"
-      class="documents-grid-header"
-    >
-      <h1 class="documents-grid-title">
-        Принятые / необработанные
-      </h1>
-    </div>
+    <DocsToVerifyList
+      :documents="filterDocs(uncheckedDocuments)"
+      verify-status="UNCHECKED"
+      @update="loadCompetencies"
+    />
 
-    <!-- Список документов -->
-    <div
-      v-if="filteredDocuments.filter((doc) => doc.verified !== false).length > 0"
-      class="documents-grid"
-    >
-      <div
-        v-for="document in filteredDocuments.filter((doc) => doc.verified !== false)"
-        :key="document.id"
-        class="document-card"
-      >
-        <div class="document-header">
-          <div class="document-icon">
-            <i class="pi pi-file" />
-          </div>
-          <div class="document-info">
-            <h3 class="document-name">
-              Документ №{{ document.id }}
-            </h3>
-          </div>
-          <div class="document-actions">
-            <Button
-              v-tooltip="'Просмотреть'"
-              icon="pi pi-eye"
-              style="background: white;"
-              class="p-button-text p-button-sm"
-              @click="viewDocument(document)"
-            />
-            <Button
-              v-tooltip="'Скачать'"
-              icon="pi pi-download"
-              style="background: white"
-              class="p-button-text p-button-sm"
-              @click="downloadDocument(document)"
-            />
-            <Button
-              v-tooltip="'Удалить'"
-              icon="pi pi-trash"
-              style="background: white"
-              class="p-button-text p-button-sm p-button-danger"
-              @click="deleteDocument(document)"
-            />
-          </div>
-        </div>
+    <DocsToVerifyList
+      :documents="filterDocs(acceptedDocuments)"
+      verify-status="ACCEPTED"
+      @update="loadCompetencies"
+    />
 
-        <div class="document-content">
-          <div class="document-details">
-            <div class="detail-item">
-              <span class="detail-label">Тип:</span>
-              <span class="detail-value">{{
-                DocumentTypes.find((type) => type.value === document.documentType)?.label
-              }}</span>
-            </div>
-            <div class="detail-item">
-              <span class="detail-label">Возрастная группа:</span>
-              <span class="detail-value">
-                {{
-                  ageGroups.find(group => document.ageCategory === group.value)?.label
-                }}</span>
-            </div>
-            <div class="detail-item">
-              <span class="detail-label">Дата загрузки:</span>
-              <span class="detail-value">{{
-                document.createdAt.substring(0, 10)
-              }}</span>
-            </div>
-            <div
-              v-if="document"
-              class="detail-item"
-            >
-              <span class="detail-label">Компетенция:</span>
-              <span class="detail-value">{{
-                documentCompetence(document)?.name
-              }}</span>
-            </div>
-          </div>
-
-          <div
-            v-if="documentExpert(document)"
-            class="mentor-info"
-          >
-            <h4 class="mentor-title">
-              Связанный эксперт:
-            </h4>
-            <p class="mentor-name">
-              {{
-                documentExpert(document)?.lastName +
-                  " " +
-                  documentExpert(document)?.firstName +
-                  " " +
-                  documentExpert(document)?.patronymic
-              }}
-            </p>
-          </div>
-          <div
-            v-if="document.verified === null"
-            class="verify"
-          >
-            <Button
-              icon="pi pi-check"
-              style="background: white"
-              label="Принять"
-              class="p-button-text p-button-sm p-button-success"
-              @click="verifyDocument(document, true)"
-            />
-            <Button
-              icon="pi pi-times"
-              style="background: white"
-              label="Отклонить"
-              class="p-button-text p-button-sm p-button-danger"
-              @click="verifyDocument(document, false)"
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div
-      v-if="filteredDocuments.filter((doc) => doc.verified === false).length > 0"
-      class="documents-grid-header"
-    >
-      <h1 class="documents-grid-title">
-        Отклоненные
-      </h1>
-    </div>
-
-    <div
-      v-if="filteredDocuments.filter((doc) => doc.verified === false).length > 0"
-      class="documents-grid"
-    >
-      <div
-        v-for="document in filteredDocuments.filter((doc) => doc.verified === false)"
-        :key="document.id"
-        class="document-card"
-      >
-        <div class="document-header">
-          <div class="document-icon">
-            <i class="pi pi-file" />
-          </div>
-          <div class="document-info">
-            <h3 class="document-name">
-              Документ №{{ document.id }}
-            </h3>
-          </div>
-          <div class="document-actions">
-            <Button
-              v-tooltip="'Просмотреть'"
-              icon="pi pi-eye"
-              style="background: white;"
-              class="p-button-text p-button-sm"
-              @click="viewDocument(document)"
-            />
-            <Button
-              v-tooltip="'Скачать'"
-              icon="pi pi-download"
-              style="background: white"
-              class="p-button-text p-button-sm"
-              @click="downloadDocument(document)"
-            />
-            <Button
-              v-tooltip="'Удалить'"
-              icon="pi pi-trash"
-              style="background: white"
-              class="p-button-text p-button-sm p-button-danger"
-              @click="deleteDocument(document)"
-            />
-          </div>
-        </div>
-
-        <div class="document-content">
-          <div class="document-details">
-            <div class="detail-item">
-              <span class="detail-label">Тип:</span>
-              <span class="detail-value">{{
-                DocumentTypes.find((type) => type.value === document.documentType)?.label
-              }}</span>
-            </div>
-            <div class="detail-item">
-              <span class="detail-label">Возрастная группа:</span>
-              <span class="detail-value">
-                {{
-                  ageGroups.find(group => document.ageCategory === group.value)?.label
-                }}</span>
-            </div>
-            <div class="detail-item">
-              <span class="detail-label">Дата загрузки:</span>
-              <span class="detail-value">{{
-                document.createdAt.substring(0, 10)
-              }}</span>
-            </div>
-            <div
-              v-if="document"
-              class="detail-item"
-            >
-              <span class="detail-label">Компетенция:</span>
-              <span class="detail-value">{{
-                documentCompetence(document)?.name
-              }}</span>
-            </div>
-          </div>
-
-          <div
-            v-if="documentExpert(document)"
-            class="mentor-info"
-          >
-            <h4 class="mentor-title">
-              Связанный эксперт:
-            </h4>
-            <p class="mentor-name">
-              {{
-                documentExpert(document)?.lastName +
-                  " " +
-                  documentExpert(document)?.firstName +
-                  " " +
-                  documentExpert(document)?.patronymic
-              }}
-            </p>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <Dialog
-      v-if="selectedDocument"
-      v-model:visible="showPreviewDialog"
-      :header="'Документ №' + selectedDocument.id"
-      :modal="true"
-      :style="{ width: '800px' }"
-    >
-      <template #footer>
-        <Button
-          label="Закрыть"
-          icon="pi pi-times"
-          class="p-button-text"
-          @click="showPreviewDialog = false"
-        />
-        <!--        <Button -->
-        <!--          label="Участники" -->
-        <!--          icon="pi pi-users" -->
-        <!--          class="p-button-primary"-->
-        <!--          @click="goToParticipants(selectedCompetence?.id)"-->
-        <!--        />-->
-      </template>
-    </Dialog>
+    <DocsToVerifyList
+      :documents="filterDocs(rejectedDocuments)"
+      verify-status="REJECTED"
+      @update="loadCompetencies"
+    />
 
     <ToastPopup :content="errors.toastPopup" />
   </div>
@@ -327,7 +98,6 @@
 
 <script lang="ts">
   import Button from "primevue/button";
-  import Dialog from "primevue/dialog";
   import Dropdown from "primevue/dropdown";
   import { FileResolver, FileType } from "@/api/resolvers/files/file.resolver";
   import { AgeCategories, CompetenceResolver } from '@/api/resolvers/competence/competence.resolver';
@@ -338,31 +108,24 @@
   import { CompetenceDocumentsResolver } from "@/api/resolvers/competenceDocuments/competence-documents.resolver";
   import type { UserOutputDto } from '@/api/resolvers/user/dto/output/user-output.dto.ts';
   import type { DocumentsOutputDto } from '@/api/resolvers/competence/dto/output/documents-output.dto.ts';
-  import apiConf from '@/api/api.conf.ts';
   import { useDocumentTemplates } from '@/shared/UseDocumentTemplates.ts';
   import { useDocumentTypes } from '@/shared/UseDocumentTypes.ts';
   import DocumentsTemplates from '@/components/DocumentsTemplates.vue';
   import { useAgeGroups } from '@/shared/UseAgeGroups.ts';
+  import DocsToVerifyList from '@/components/DocsToVerifyList.vue';
 
   export default {
     name: "AdminDocuments",
     components: {
+      DocsToVerifyList,
       DocumentsTemplates,
       ToastPopup,
       Button,
-      Dialog,
       Dropdown,
     },
     data() {
       return {
         ageGroups: useAgeGroups,
-        fileResolver: new FileResolver(),
-        filePreview: null,
-        competenceResolver: new CompetenceResolver(),
-        userResolver: new UserResolver(),
-        showUploadDialog: false,
-        showLinkDialog: false,
-        selectedDocument: null as CompetenceDocumentsOutputDto | null,
         selectedType: null as null | FileType,
         selectedCompetence: localStorage.getItem("selectedCompetence")
           ? JSON.parse(localStorage.getItem("selectedCompetence") as string)
@@ -379,31 +142,21 @@
           },
         },
         selectedAge: null as null | AgeCategories,
-        showPreviewDialog: false,
+        fileResolver: new FileResolver(),
+        competenceResolver: new CompetenceResolver(),
+        userResolver: new UserResolver(),
         competenceDocumentsResolver: new CompetenceDocumentsResolver(),
       };
     },
     computed: {
-      filteredDocuments() {
-        let filtered = this.documents;
-
-        if (this.selectedType) {
-          filtered = filtered.filter(
-            (doc) => doc.documentType === this.selectedType,
-          );
-        }
-
-        if (this.selectedCompetence) {
-          filtered = filtered.filter(
-            (doc) => this.selectedCompetence === this.documentCompetence(doc),
-          );
-        }
-
-        if (this.selectedAge) {
-          filtered = filtered.filter((d) => d.ageCategory === this.selectedAge);
-        }
-
-        return filtered;
+      rejectedDocuments() {
+        return this.documents.filter(doc => doc.verified === false)
+      },
+      acceptedDocuments() {
+        return this.documents.filter(doc => doc.verified === true)
+      },
+      uncheckedDocuments() {
+        return this.documents.filter(doc => doc.verified === null)
       },
       availableAges() {
         const ageOrder = new Map(this.ageGroups.map((group, index) => [group.value, index]));
@@ -421,40 +174,36 @@
           competence.documents.some((doc) => doc.id === document.id),
         );
       },
-      documentExpert(document: CompetenceDocumentsOutputDto) {
-        return this.experts.find((expert) => expert.id === document.userId);
-      },
-      async verifyDocument(doc: CompetenceDocumentsOutputDto, status: boolean) {
-        const response = await this.competenceDocumentsResolver.verify(doc.id, status)
-        if (response.status === 200) await this.loadCompetencies()
-      },
-      viewDocument(doc: CompetenceDocumentsOutputDto) {
-        this.selectedDocument = doc;
-        window.open(`${apiConf.endpoint}/file-service/v1/files/view/${doc.documentId}`, "_blank");
-      },
-      downloadDocument(doc: CompetenceDocumentsOutputDto) {
-        window.location.href = `${apiConf.endpoint}/file-service/v1/files/download/${doc.documentId}`;
-      },
-      async deleteDocument(document: CompetenceDocumentsOutputDto) {
-        if (
-          confirm(`Вы уверены, что хотите удалить документ "${document.documentId}"?`)
-        ) {
-          const response = await this.competenceDocumentsResolver.delete(
-            document.id,
+      filterDocs(docs: CompetenceDocumentsOutputDto[]) {
+        if (this.selectedType) {
+          docs = docs.filter(
+            (doc) => doc.documentType === this.selectedType,
           );
-          if (response.status === 200) {
-            this.competencies = [];
-            this.documents = [];
-            this.experts = [];
-            await this.loadCompetencies();
-          }
         }
+
+        if (this.selectedCompetence) {
+          docs = docs.filter(
+            (doc) => this.selectedCompetence === this.documentCompetence(doc),
+          );
+        }
+
+        if (this.selectedAge) {
+          docs = docs.filter((d) => d.ageCategory === this.selectedAge);
+        }
+
+        return docs;
       },
       resetFilters() {
         this.selectedType = null;
         this.selectedCompetence = null;
       },
+      resetAge() {
+        this.selectedAge = null
+      },
       async loadCompetencies() {
+        this.documents = []
+        this.experts = []
+        this.competencies = []
         const response = await this.competenceResolver.getAll();
         if (response.status === 200 && typeof response.message !== "string") {
           response.message.forEach((competence) => {
@@ -511,7 +260,7 @@
     }
   }
 
-  .page-header, .documents-grid-header {
+  .page-header {
     margin-bottom: 2rem;
   }
 
@@ -563,166 +312,6 @@
   .filter-dropdown {
     width: 100%;
   }
-
-  .documents-grid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(400px, 1fr));
-    gap: 1.5rem;
-    width: 100%;
-    margin-bottom: 4rem;
-  }
-
-  .documents-grid-title {
-    color: #2c3e50;
-    margin: 0 0 0.5rem 0;
-    font-size: 1.5rem;
-    font-weight: 500;
-    font-family: "BIPS", sans-serif;
-  }
-
-  .document-card {
-    background: white;
-    border-radius: 12px;
-    box-shadow: 0 2px 12px rgba(0, 0, 0, 0.08);
-    border: 2px solid transparent;
-    overflow: hidden;
-    transition:
-      box-shadow 0.3s ease,
-      border-color 0.3s ease;
-  }
-
-  .document-card:hover {
-    box-shadow: 0 4px 20px rgba(255, 152, 0, 0.2);
-    border: 2px solid #ff9800;
-  }
-
-  .document-header {
-    background: linear-gradient(135deg, #ff9800, #f57c00);
-    color: white;
-    padding: 1.5rem;
-    display: flex;
-    align-items: center;
-    gap: 1rem;
-  }
-
-  .document-icon {
-    width: 60px;
-    height: 60px;
-    background: rgba(255, 255, 255, 0.2);
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 1.5rem;
-    flex-shrink: 0;
-  }
-
-  .document-info {
-    flex: 1;
-  }
-
-  .document-name {
-    margin: 0 0 0.25rem 0;
-    font-size: 1.1rem;
-    font-weight: 600;
-  }
-
-  .document-type {
-    margin: 0;
-    font-size: 0.9rem;
-    opacity: 0.9;
-  }
-
-  .document-actions {
-    display: flex;
-    gap: 0.5rem;
-  }
-
-  .document-content {
-    padding: 1.5rem;
-  }
-
-  .document-details {
-    margin-bottom: 1.5rem;
-  }
-
-  .detail-item {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 0.5rem 0;
-    border-bottom: 1px solid #f1f3f4;
-  }
-
-  .detail-item:last-child {
-    border-bottom: none;
-  }
-
-  .detail-label {
-    color: #6c757d;
-    font-weight: 500;
-    min-width: 100px;
-  }
-
-  .detail-value {
-    color: #2c3e50;
-    font-weight: 500;
-    text-align: right;
-  }
-
-  .status-badge {
-    padding: 0.25rem 0.75rem;
-    border-radius: 20px;
-    font-size: 0.8rem;
-    font-weight: 500;
-    text-transform: uppercase;
-    letter-spacing: 0.5px;
-  }
-
-  .status-verified {
-    background: rgba(40, 167, 69, 0.2);
-    color: #28a745;
-    border: 1px solid #28a745;
-  }
-
-  .status-pending {
-    background: rgba(255, 193, 7, 0.2);
-    color: #ffc107;
-    border: 1px solid #ffc107;
-  }
-
-  .status-rejected {
-    background: rgba(220, 53, 69, 0.2);
-    color: #dc3545;
-    border: 1px solid #dc3545;
-  }
-
-  .expert-info {
-    background: #f8f9fa;
-    padding: 1rem;
-    border-radius: 8px;
-    border-left: 4px solid #ff9800;
-  }
-
-  .expert-title {
-    color: #2c3e50;
-    margin: 0 0 0.5rem 0;
-    font-size: 0.9rem;
-    font-weight: 600;
-  }
-
-  .expert-name {
-    color: #6c757d;
-    margin: 0;
-    font-size: 0.9rem;
-  }
-
-  .verify {
-    display: flex;
-    gap: 1.5rem;
-    justify-content: flex-end;
-  }
-
 
   /* Формы */
   .upload-form,
